+++ src/components/custom-field-input-currency/custom-field-input-currency.jsx
@@ -76,7 +76,12 @@ const CustomFieldInputCurrency = forwardRef(function CustomFieldInputCurrency(pr
   });
 
   useImperativeHandle(ref, () => ({
+    get dirty() {
+      const currentValue = this.value ? this.value[0] : this.value;
+      return props.value !== currentValue;
+    },
     id: props.id,
+    name: props.name,
     get value() {
       let numberValue;
 
@@ -140,7 +145,7 @@ CustomFieldInputCurrency.propTypes = {
   errorText: PropTypes.string,
   id: PropTypes.string.isRequired,
   label: PropTypes.string.isRequired,
-  name: PropTypes.string,
+  name: PropTypes.string.isRequired,
   // onChange: Do not expose an onChange handler. See commit for details.
   placeholder: PropTypes.string,
   readOnly: PropTypes.bool,