+++ src/components/custom-field-input-currency/custom-field-input-currency.test.jsx
@@ -5,32 +5,35 @@ import {
   screen,
 } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import renderer from 'react-test-renderer';
 import CustomFieldInputText from '../custom-field-input-text/custom-field-input-text.jsx';
 import CustomFieldInputCurrency from './custom-field-input-currency.jsx';
 
 describe('CustomFieldInputCurrency', () => {
-  const renderComponent = (props = {}) => render(<CustomFieldInputCurrency id="currency" label="currency" {...props} />);
+  const requiredProps = {
+    id: 'currency',
+    label: 'currency',
+    name: 'some-field-id',
+  };
 
   it('has defaults', () => {
-    const tree = renderer.create((
-      <CustomFieldInputCurrency id="foo" label="cash money" />
-    )).toJSON();
-    expect(tree).toMatchSnapshot();
+    const ref = createRef();
+    render(<CustomFieldInputCurrency {...requiredProps} ref={ref} />);
+    expect(document.body).toMatchSnapshot();
+    expect(ref.current).toMatchSnapshot();
   });
 
   it('accepts a currency code', () => {
-    const { getByLabelText } = renderComponent({ value: 5000, currencyCode: 'XAF' });
-    expect(getByLabelText('currency').value).toMatch(/FCFA/);
+    render(<CustomFieldInputCurrency {...requiredProps} value={5000} currencyCode="XAF" />);
+    expect(screen.getByLabelText('currency').value).toMatch(/FCFA/);
   });
 
   it('sets the correct step according to the provided currency code', () => {
-    const { getByLabelText } = renderComponent({ currencyCode: 'IQD', value: 10111 });
+    render(<CustomFieldInputCurrency {...requiredProps} currencyCode="IQD" value={10111} />);
 
     // NOTE: This space is character code 160, non-breaking space. It did break my brain though
-    expect(getByLabelText('currency')).toHaveValue('IQD 10.111');
-    fireEvent.focus(getByLabelText('currency'));
-    expect(getByLabelText('currency')).toHaveValue(10.111);
+    expect(screen.getByLabelText('currency')).toHaveValue('IQD 10.111');
+    fireEvent.focus(screen.getByLabelText('currency'));
+    expect(screen.getByLabelText('currency')).toHaveValue(10.111);
   });
 
   describe('prop-forward API', () => {
@@ -47,78 +50,75 @@ describe('CustomFieldInputCurrency', () => {
 
     it('presents contextual error state', () => {
       const errorText = 'What do you want from us monster!?';
-      renderComponent({ value: 350, errorText, error: true });
+      render(<CustomFieldInputCurrency {...requiredProps} value={350} errorText={errorText} error={true} />);
 
       expect(screen.getByLabelText('currency')).toBeInvalid();
       expect(screen.getByText(errorText)).toBeInTheDocument();
     });
+  });
 
-    it('renders all forwarded props except event handlers', () => {
-      const tree = renderer.create((
-        <CustomFieldInputCurrency
-          className="test"
-          disabled={false}
-          id="cFa-id"
-          label="cFa-label"
-          name="cFa-name"
-          placeholder="cFa-placeholder"
-          required={true}
-          value={10}
-        />
-      )).toJSON();
-      const stringTree = JSON.stringify(tree);
-
-      expect(tree.props.className).toContain('test');
-      expect(tree.props.className).not.toContain('disabled');
-
-      expect(stringTree).toContain('cFa-label');
-      expect(stringTree).toContain('cFa-id');
-      expect(stringTree).toContain('cFa-name');
-      expect(stringTree).toContain('cFa-placeholder');
-      expect(stringTree).toContain('Required');
-      expect(stringTree).toContain('10');
+  describe('dirty ref API', () => {
+    it('updates on user interactions', () => {
+      const ref = createRef();
+      render(<CustomFieldInputCurrency {...requiredProps} ref={ref} />);
+      userEvent.click(screen.getByLabelText('currency'));
+      expect(ref.current.dirty).toEqual(false);
+      userEvent.type(screen.getByLabelText('currency'), '12');
+      expect(ref.current.dirty).toEqual(true);
+      userEvent.type(screen.getByLabelText('currency'), '{backspace}');
+      expect(ref.current.dirty).toEqual(true);
+      userEvent.type(screen.getByLabelText('currency'), '{backspace}');
+      expect(ref.current.dirty).toEqual(false);
     });
   });
 
   describe('disabled API', () => {
     it('respects the disabled prop', () => {
-      const { getByLabelText } = renderComponent({ disabled: true });
-      expect(getByLabelText('currency')).toBeDisabled();
+      render(<CustomFieldInputCurrency {...requiredProps} disabled={true} />);
+      expect(screen.getByLabelText('currency')).toBeDisabled();
     });
 
     it('does not enter edit mode on focus', () => {
-      const { getByLabelText } = renderComponent({ disabled: true });
-      fireEvent.focus(getByLabelText('currency'));
-      expect(getByLabelText('currency')).toHaveAttribute('type', 'text');
+      render(<CustomFieldInputCurrency {...requiredProps} disabled={true} />);
+      fireEvent.focus(screen.getByLabelText('currency'));
+      expect(screen.getByLabelText('currency')).toHaveAttribute('type', 'text');
     });
 
     it('respects the enabled prop', () => {
-      const { getByLabelText } = renderComponent({ disabled: false });
-      expect(getByLabelText('currency')).not.toBeDisabled();
+      render(<CustomFieldInputCurrency {...requiredProps} disabled={false} />);
+      expect(screen.getByLabelText('currency')).not.toBeDisabled();
+    });
+  });
+
+  describe('name ref API', () => {
+    it('can be set', () => {
+      const ref = createRef();
+      render(<CustomFieldInputCurrency {...requiredProps} name="some-unique-name" ref={ref} />);
+      expect(ref.current.name).toEqual('some-unique-name');
     });
   });
 
   describe('readOnly API', () => {
     it('respects the readOnly prop', () => {
-      const { getByLabelText } = renderComponent({ readOnly: true });
-      expect(getByLabelText('currency')).toHaveAttribute('readOnly', '');
+      render(<CustomFieldInputCurrency {...requiredProps} readOnly={true} />);
+      expect(screen.getByLabelText('currency')).toHaveAttribute('readOnly', '');
     });
 
     it('does not enter edit mode on focus', () => {
-      const { getByLabelText } = renderComponent({ readOnly: true });
-      fireEvent.focus(getByLabelText('currency'));
-      expect(getByLabelText('currency')).toHaveAttribute('type', 'text');
+      render(<CustomFieldInputCurrency {...requiredProps} readOnly={true} />);
+      fireEvent.focus(screen.getByLabelText('currency'));
+      expect(screen.getByLabelText('currency')).toHaveAttribute('type', 'text');
     });
 
     it('is false by default', () => {
-      const { getByLabelText } = renderComponent();
-      expect(getByLabelText('currency')).not.toHaveAttribute('readOnly', '');
+      render(<CustomFieldInputCurrency {...requiredProps} />);
+      expect(screen.getByLabelText('currency')).not.toHaveAttribute('readOnly', '');
     });
   });
 
   describe('input validation', () => {
     it('does not trigger error state for valid value', () => {
-      renderComponent();
+      render(<CustomFieldInputCurrency {...requiredProps} />);
 
       fireEvent.focus(screen.getByLabelText('currency'));
       fireEvent.change(screen.getByLabelText('currency'), { target: { value: 1234 } });
@@ -129,39 +129,39 @@ describe('CustomFieldInputCurrency', () => {
     });
 
     it('does not switch to view mode when its value is numerically invalid', () => {
-      const { getByLabelText } = renderComponent();
+      render(<CustomFieldInputCurrency {...requiredProps} />);
 
-      fireEvent.focus(getByLabelText('currency'));
-      fireEvent.change(getByLabelText('currency'), { target: { value: 12.111111 } });
-      fireEvent.blur(getByLabelText('currency'));
+      fireEvent.focus(screen.getByLabelText('currency'));
+      fireEvent.change(screen.getByLabelText('currency'), { target: { value: 12.111111 } });
+      fireEvent.blur(screen.getByLabelText('currency'));
 
-      expect(getByLabelText('currency')).toHaveAttribute('type', 'number');
-      expect(getByLabelText('currency')).toHaveValue(12.111111);
+      expect(screen.getByLabelText('currency')).toHaveAttribute('type', 'number');
+      expect(screen.getByLabelText('currency')).toHaveValue(12.111111);
     });
 
     xit('does not switch to view mode when its value is numerically invalid', () => {
       // This test does not work because programmatically setting value
       // on a number input does not reflect a user typing.
       // See commit for more details.
-      const { getByLabelText } = renderComponent();
+      render(<CustomFieldInputCurrency {...requiredProps} />);
 
-      fireEvent.focus(getByLabelText('currency'));
-      fireEvent.change(getByLabelText('currency'), { target: { value: '12..' } });
-      fireEvent.blur(getByLabelText('currency'));
+      fireEvent.focus(screen.getByLabelText('currency'));
+      fireEvent.change(screen.getByLabelText('currency'), { target: { value: '12..' } });
+      fireEvent.blur(screen.getByLabelText('currency'));
 
-      expect(getByLabelText('currency')).toHaveAttribute('type', 'number');
-      expect(getByLabelText('currency')).toHaveValue('12..');
+      expect(screen.getByLabelText('currency')).toHaveAttribute('type', 'number');
+      expect(screen.getByLabelText('currency')).toHaveValue('12..');
     });
 
     it('accepts an undefined value', () => {
-      const { getByLabelText } = renderComponent({ value: undefined });
-      expect(getByLabelText('currency')).toHaveValue('');
+      render(<CustomFieldInputCurrency {...requiredProps} value={undefined} />);
+      expect(screen.getByLabelText('currency')).toHaveValue('');
     });
   });
 
   describe('displayed value', () => {
     it('does not try to parse an empty input to float', () => {
-      renderComponent();
+      render(<CustomFieldInputCurrency {...requiredProps} />);
 
       userEvent.click(screen.getByLabelText('currency'));
       userEvent.click(document.body);
@@ -173,21 +173,21 @@ describe('CustomFieldInputCurrency', () => {
 
   describe('value API', () => {
     it('can be undefined', () => {
-      renderComponent();
+      render(<CustomFieldInputCurrency {...requiredProps} />);
       expect(screen.getByLabelText('currency')).toHaveValue('');
       userEvent.click(screen.getByLabelText('currency'));
       expect(screen.getByLabelText('currency')).toHaveValue(null);
     });
 
     it('can be 0', () => {
-      renderComponent({ value: 0 });
+      render(<CustomFieldInputCurrency {...requiredProps} value={0} />);
       expect(screen.getByLabelText('currency')).toHaveValue('$0.00');
       userEvent.click(screen.getByLabelText('currency'));
       expect(screen.getByLabelText('currency')).toHaveValue(0);
     });
 
     it('can be a positive integer', () => {
-      renderComponent({ value: 1 });
+      render(<CustomFieldInputCurrency {...requiredProps} value={1} />);
       expect(screen.getByLabelText('currency')).toHaveValue('$0.01');
       userEvent.click(screen.getByLabelText('currency'));
       expect(screen.getByLabelText('currency')).toHaveValue(0.01);
@@ -196,37 +196,37 @@ describe('CustomFieldInputCurrency', () => {
 
   describe('value ref API', () => {
     it('can get a positive value', () => {
-      const inputRef = createRef(null);
-      render(<CustomFieldInputCurrency id="test-input" label="Test label" ref={inputRef} />);
+      const ref = createRef(null);
+      render(<CustomFieldInputCurrency {...requiredProps} id="test-input" label="Test label" ref={ref} />);
 
       fireEvent.focus(screen.getByLabelText('Test label'));
       fireEvent.change(screen.getByLabelText('Test label'), { target: { value: 1234 } });
-      expect(inputRef.current.value).toStrictEqual([123400, 'USD']);
+      expect(ref.current.value).toStrictEqual([123400, 'USD']);
       fireEvent.blur(screen.getByLabelText('Test label'));
-      expect(inputRef.current.value).toStrictEqual([123400, 'USD']);
+      expect(ref.current.value).toStrictEqual([123400, 'USD']);
     });
 
     it('can get a negative value', () => {
-      const inputRef = createRef(null);
-      render(<CustomFieldInputCurrency id="test-input" label="Test label" ref={inputRef} />);
+      const ref = createRef(null);
+      render(<CustomFieldInputCurrency {...requiredProps} id="test-input" label="Test label" ref={ref} />);
 
       fireEvent.focus(screen.getByLabelText('Test label'));
       fireEvent.change(screen.getByLabelText('Test label'), { target: { value: -1234 } });
-      expect(inputRef.current.value).toStrictEqual([-123400, 'USD']);
+      expect(ref.current.value).toStrictEqual([-123400, 'USD']);
       fireEvent.blur(screen.getByLabelText('Test label'));
-      expect(inputRef.current.value).toStrictEqual([-123400, 'USD']);
+      expect(ref.current.value).toStrictEqual([-123400, 'USD']);
     });
 
     it('can be undefined', () => {
-      const inputRef = createRef(null);
-      render(<CustomFieldInputCurrency id="test-input" label="Test label" ref={inputRef} />);
+      const ref = createRef(null);
+      render(<CustomFieldInputCurrency {...requiredProps} id="test-input" label="Test label" ref={ref} />);
 
       fireEvent.focus(screen.getByLabelText('Test label'));
       fireEvent.change(screen.getByLabelText('Test label'), { target: { value: '' } });
 
-      expect(inputRef.current.value).toBeUndefined();
+      expect(ref.current.value).toBeUndefined();
       fireEvent.blur(screen.getByLabelText('Test label'));
-      expect(inputRef.current.value).toBeUndefined();
+      expect(ref.current.value).toBeUndefined();
     });
   });
 });