+++ src/components/custom-field-input-date/custom-field-input-date.test.jsx
@@ -1,54 +1,70 @@
 import React, { createRef } from 'react';
 import {
-  cleanup,
   fireEvent,
   render,
   screen,
 } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import renderer from 'react-test-renderer';
 import CustomFieldInputDate from './custom-field-input-date.jsx';
 
 describe('src/components/custom-field-input-date/custom-field-input-date', () => {
-  const renderComponent = (props = {}) => render(<CustomFieldInputDate label="Field Date" id="field-date" {...props} />);
+  const requiredProps = {
+    id: 'field-date',
+    label: 'Field Date',
+    name: 'field-id',
+  };
+
   const changeValue = (getInputElement, value) => {
     userEvent.click(getInputElement());
     fireEvent.change(getInputElement(), { target: { value } });
     fireEvent.blur(getInputElement());
   };
 
-  afterEach(cleanup);
-
   it('has defaults', () => {
-    const tree = renderer.create(<CustomFieldInputDate label="Hello" id="hello" />).toJSON();
-    expect(tree).toMatchSnapshot();
+    const ref = createRef();
+    render(<CustomFieldInputDate {...requiredProps} ref={ref} />);
+    expect(document.body).toMatchSnapshot();
+    expect(ref.current).toMatchSnapshot();
+  });
+
+  describe('dirty ref API', () => {
+    xit('updates on user interactions', () => {
+      const ref = createRef();
+      render(<CustomFieldInputDate {...requiredProps} ref={ref} />);
+      userEvent.click(screen.getByLabelText('Field Date'));
+      expect(ref.current.dirty).toEqual(false);
+      userEvent.type(screen.getByLabelText('Field Date'), '12/12/1212');
+      expect(ref.current.dirty).toEqual(true);
+      userEvent.type(screen.getByLabelText('Field Date'), '');
+      expect(ref.current.dirty).toEqual(false);
+    });
   });
 
   describe('value API', () => {
     it('accepts a string in format YYYY-MM-DD', () => {
-      const { getByLabelText } = renderComponent({ value: '2016-07-18' });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="2016-07-18" />);
       expect(getByLabelText('Field Date')).toHaveValue('July 18, 2016');
     });
 
     it('accepts a string in format mm/dd/yyyy', () => {
-      const { getByLabelText } = renderComponent({ value: '07/18/2016' });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="07/18/2016" />);
       expect(getByLabelText('Field Date')).toHaveValue('July 18, 2016');
     });
 
     it('accepts a string in format mm-dd-yyyy', () => {
-      const { getByLabelText } = renderComponent({ value: '07-18-2016' });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="07-18-2016" />);
       expect(getByLabelText('Field Date')).toHaveValue('July 18, 2016');
     });
   });
 
   describe('disabled API', () => {
     it('permits itself to be disabled', () => {
-      const { getByLabelText } = renderComponent({ disabled: true });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} disabled={true} />);
       expect(getByLabelText('Field Date')).toBeDisabled();
     });
 
     it('permits itself to be enabled', () => {
-      const { getByLabelText } = renderComponent({ disabled: false });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} disabled={false} />);
       expect(getByLabelText('Field Date')).not.toBeDisabled();
     });
   });
@@ -56,7 +72,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
   describe('error API', () => {
     describe('when the value is valid', () => {
       it('does not show an error', () => {
-        render(<CustomFieldInputDate label="Field Date" id="field-date" value="05/10/1992" />);
+        render(<CustomFieldInputDate {...requiredProps} label="Field Date" id="field-date" value="05/10/1992" />);
         expect(screen.getByLabelText('Field Date')).toBeValid('error');
       });
     });
@@ -64,7 +80,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
 
   describe('errorText', () => {
     it('shows an error icon and the error text', () => {
-      renderComponent({ value: '2016-07-18', errorText: 'Bad Date!' });
+      render(<CustomFieldInputDate {...requiredProps} value="2016-07-18" errorText="Bad Date!" />);
       expect(screen.getByLabelText('Field Date')).not.toBeValid();
       expect(screen.getByText('Bad Date!')).toBeInTheDocument();
     });
@@ -72,12 +88,12 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
 
   describe('required API', () => {
     it('is required when told to be', () => {
-      const { getByLabelText } = renderComponent({ required: true });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} required={true} />);
       expect(getByLabelText('Field Date')).toBeRequired();
     });
 
     it('is not required when told not to be', () => {
-      const { getByLabelText } = renderComponent({ required: false });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} required={false} />);
       expect(getByLabelText('Field Date')).not.toBeRequired();
     });
   });
@@ -85,7 +101,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
   describe('interaction', () => {
     describe('when focused', () => {
       it('switches to a date input when focused and opens the calendar', () => {
-        const { getByLabelText } = renderComponent({ value: '07/18/2016' });
+        const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="07/18/2016" />);
         expect(getByLabelText('Field Date')).toHaveAttribute('type', 'text');
         userEvent.click(getByLabelText('Field Date'));
         expect(getByLabelText('Field Date')).toHaveAttribute('type', 'date');
@@ -95,14 +111,14 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
 
     describe('when clicked', () => {
       it('ensures the date input shows the correct value', () => {
-        const { getByLabelText } = renderComponent({ value: '07/18/2016' });
+        const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="07/18/2016" />);
         expect(getByLabelText('Field Date')).toHaveAttribute('type', 'text');
         userEvent.click(getByLabelText('Field Date'));
         expect(getByLabelText('Field Date')).toHaveValue('2016-07-18');
       });
 
       it('focuses on the date input', () => {
-        const { getByLabelText } = renderComponent({ value: '07/18/2016' });
+        const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="07/18/2016" />);
         expect(getByLabelText('Field Date')).toHaveAttribute('type', 'text');
         userEvent.click(getByLabelText('Field Date'));
         expect(getByLabelText('Field Date')).toHaveFocus();
@@ -113,7 +129,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
       it('switches over to the text input', () => {
         render(
           <div>
-            <CustomFieldInputDate label="Field Date" id="field-date" value={'07/18/2016'} />
+            <CustomFieldInputDate {...requiredProps} value={'07/18/2016'} />
             <input aria-label={'force-blur'} />
           </div>,
         );
@@ -125,7 +141,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
       });
 
       it('stays in edit mode when it is invalid', () => {
-        const { getByLabelText } = renderComponent({ value: '07/18/2016', error: true });
+        const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="07/18/2016" error />);
         userEvent.click(getByLabelText('Field Date'));
         fireEvent.blur(getByLabelText('Field Date'));
         expect(getByLabelText('Field Date')).toHaveAttribute('type', 'date');
@@ -135,7 +151,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
 
   describe('calendar icon', () => {
     it('opens to the correct date on click and closes when clicking away', () => {
-      renderComponent({ value: '2016-07-18' });
+      render(<CustomFieldInputDate {...requiredProps} value="2016-07-18" />);
       expect(screen.queryByText('July 2016')).not.toBeInTheDocument();
       userEvent.click(screen.getByTitle('Field Date calendar button'));
       expect(screen.getByText('July 2016')).toBeInTheDocument();
@@ -144,7 +160,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
     });
 
     it('opens to the correct date on click / pressing enter and closes on escape', () => {
-      const { getByLabelText } = renderComponent({ value: '2016-07-18' });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="2016-07-18" />);
       expect(screen.queryByText('July 2016')).not.toBeInTheDocument();
       userEvent.click(getByLabelText('Field Date'));
       expect(screen.getByText('July 2016')).toBeInTheDocument();
@@ -155,7 +171,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
     });
 
     it('changes the date to the date selected', () => {
-      const { getByLabelText } = renderComponent({ value: '2016-09-13' });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="2016-09-13" />);
       expect(getByLabelText('Field Date')).toHaveValue('September 13, 2016');
       userEvent.click(screen.getByTitle('Field Date calendar button'));
       userEvent.click(screen.getByLabelText('September 14'));
@@ -163,7 +179,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
     });
 
     it('does not open when disabled', () => {
-      renderComponent({ value: '2016-07-18', disabled: true });
+      render(<CustomFieldInputDate {...requiredProps} value="2016-07-18" disabled />);
       expect(screen.queryByText('July 2016')).not.toBeInTheDocument();
       userEvent.click(screen.getByTitle('Field Date calendar'));
       expect(screen.queryByText('July 2016')).not.toBeInTheDocument();
@@ -173,7 +189,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
   describe('onChange API', () => {
     it('accepts an onChange listener', () => {
       const onChange = jest.fn();
-      const { getByLabelText } = renderComponent({ value: '', onChange });
+      const { getByLabelText } = render(<CustomFieldInputDate {...requiredProps} value="" onChange={onChange} />);
       changeValue(() => getByLabelText('Field Date'), '2016-07-18');
       expect(onChange.mock.calls.length).toBe(1);
     });
@@ -181,7 +197,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
 
   describe('when the component is readOnly', () => {
     it('does not display the helpText', () => {
-      renderComponent({ readOnly: true, helpText: 'Foo', error: true });
+      render(<CustomFieldInputDate {...requiredProps} readOnly={true} helpText="Foo" error={true} />);
       expect(screen.queryByText('Foo')).not.toBeInTheDocument();
     });
   });
@@ -189,7 +205,7 @@ describe('src/components/custom-field-input-date/custom-field-input-date', () =>
   describe('forwardRef API', () => {
     it('can be used to get value', () => {
       const inputRef = createRef(null);
-      render(<CustomFieldInputDate id="test-input" label="Test label" ref={inputRef} value="2016-07-18" />);
+      render(<CustomFieldInputDate {...requiredProps} id="test-input" label="Test label" ref={inputRef} value="2016-07-18" />);
 
       expect(inputRef.current.value).toBe('July 18, 2016');
     });