+++ src/components/custom-field-input-text/custom-field-input-text.jsx
@@ -10,7 +10,12 @@ const CustomFieldInputText = forwardRef(function CustomFieldInputText(props, ref
   const validationMessage = useValidation(props.readOnly, props.errorText, inputRef);
 
   useImperativeHandle(ref, () => ({
+    get dirty() {
+      const providedValue = props.defaultValue || '';
+      return providedValue !== this.value;
+    },
     id: props.id,
+    name: props.name,
     get value() {
       return inputRef.current.value;
     },