+++ src/components/custom-field-input-text/custom-field-input-text.test.jsx
@@ -1,35 +1,51 @@
 import React, { createRef } from 'react';
-import renderer from 'react-test-renderer';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import CustomFieldInputText from './custom-field-input-text.jsx';
 
 describe('CustomFieldInputText', () => {
-  function TestComponent(props = {}) {
-    return <CustomFieldInputText id="test-input" label="Test label" {...props} />;
-  }
+  const requiredProps = {
+    id: 'test-input',
+    label: 'Test label',
+    name: 'field-id',
+  };
 
   it('has defaults', () => {
-    expect(renderer.create(<TestComponent />).toJSON()).toMatchSnapshot();
+    const ref = createRef();
+    render(<CustomFieldInputText {...requiredProps} ref={ref} />);
+    expect(document.body).toMatchSnapshot();
+    expect(ref.current).toMatchSnapshot();
+  });
+
+  describe('dirty ref API', () => {
+    it('updates on user interactions', () => {
+      const ref = createRef();
+      render(<CustomFieldInputText {...requiredProps} ref={ref} defaultValue="ab" />);
+      expect(ref.current.dirty).toEqual(false);
+      userEvent.type(screen.getByLabelText('Test label'), '{backspace}');
+      expect(ref.current.dirty).toEqual(true);
+      userEvent.type(screen.getByLabelText('Test label'), '{backspace}');
+      expect(ref.current.dirty).toEqual(true);
+    });
   });
 
   describe('errorText API', () => {
     it('can have an error state through a native validation', () => {
       // I am not sure what is the best way to represent this in a test.
       // However, at the moment, there are end-to-end tests in the Number component tests.
-      render(<TestComponent errorText="yo" />);
+      render(<CustomFieldInputText {...requiredProps} errorText="yo" />);
       expect(screen.getByLabelText('Test label')).toBeInvalid();
       expect(screen.getByText('Invalid custom field')).toBeInTheDocument();
     });
 
     it('can have an error state through a custom validation', () => {
-      render(<TestComponent errorText="Custom validation message" />);
+      render(<CustomFieldInputText {...requiredProps} errorText="Custom validation message" />);
       expect(screen.getByLabelText('Test label')).toBeInvalid();
       expect(screen.getByText('Invalid custom field')).toBeInTheDocument();
     });
 
     it('can have no error state', () => {
-      render(<TestComponent />);
+      render(<CustomFieldInputText {...requiredProps} />);
       expect(screen.getByLabelText('Test label')).toBeValid();
       expect(screen.queryByRole('[role="img"]')).toBeFalsy();
     });
@@ -38,14 +54,14 @@ describe('CustomFieldInputText', () => {
   describe('inputRef API', () => {
     it('sets the ref on the input', () => {
       const inputRef = createRef();
-      render(<TestComponent inputRef={inputRef} />);
+      render(<CustomFieldInputText {...requiredProps} inputRef={inputRef} />);
       expect(screen.getByLabelText('Test label')).toBe(inputRef.current);
     });
   });
 
   describe('value API', () => {
     it('sets the value attribute', () => {
-      render(<TestComponent value="test-value" />);
+      render(<CustomFieldInputText {...requiredProps} value="test-value" />);
       expect(screen.getByLabelText('Test label')).toHaveValue('test-value');
     });
   });